backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
feedparser==6.0.11
python-dateutil==2.9.0.post0
bleach==6.1.0
cryptography==41.0.7
slowapi==0.1.9
chardet==5.2.0
